Prob Project-student/requirements.txt
PyQt5>=5.15.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=0.24.0
statsmodels>=0.13.0
Pillow>=8.0.0 